Fix validation error bug
